# contracts/openapi/public-service.json
{
  "components": {
    "schemas": {
...
      "ChangeBatchDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "ids",
          "decision"
        ],
        "title": "ChangeBatchDecisionRequest",
        "type": "object"
      },
      "ChangeBatchDecisionResponse": {
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ChangeBatchDecisionItemResult"
            },
            "title": "Results",
            "type": "array"
          },
          "succeeded": {
            "title": "Succeeded",
            "type": "integer"
          },
          "total_requested": {
            "title": "Total Requested",
            "type": "integer"
          }
        },
        "required": [
          "decision",
          "total_requested",
          "succeeded",
          "failed",
          "results"
        ],
        "title": "ChangeBatchDecisionResponse",
        "type": "object"
      },
      "ChangeDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ChangeDecisionRequest",
        "type": "object"
      },
...
      "ChangeEditRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "enum": [
              "proposal",
              "canonical"
            ],
            "title": "Mode",
            "type": "string"
          },
          "patch": {
            "$ref": "#/components/schemas/ChangeEditPatchRequest"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "target": {
            "$ref": "#/components/schemas/ChangeEditTargetRequest"
          }
        },
        "required": [
          "mode",
          "target",
          "patch"
        ],
        "title": "ChangeEditRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CalendarDIFF API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/changes": {
      "get": {
        "operationId": "get_changes_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "default": "pending",
              "title": "Review Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChangeItemResponse"
                  },
                  "title": "Response Get Changes Changes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Changes",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/batch/decisions": {
      "post": {
        "operationId": "post_change_batch_decisions_changes_batch_decisions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeBatchDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeBatchDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Batch Decisions",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/edits": {
      "post": {
        "operationId": "post_change_edit_apply_changes_edits_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEditApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Edit Apply",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/edits/preview": {
      "post": {
        "operationId": "post_change_edit_preview_changes_edits_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEditPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Edit Preview",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/summary": {
      "get": {
        "operationId": "get_changes_summary_route_changes_summary_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangesWorkbenchSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Changes Summary Route",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}": {
      "get": {
        "operationId": "get_change_item_changes__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Item",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/decisions": {
      "post": {
        "operationId": "post_change_decision_changes__change_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Decision",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/edit-context": {
      "get": {
        "operationId": "get_change_edit_context_changes__change_id__edit_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEditContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Edit Context",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/evidence/{side}/preview": {
      "get": {
        "operationId": "get_change_evidence_preview_changes__change_id__evidence__side__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "side",
            "required": true,
            "schema": {
              "enum": [
                "before",
                "after"
              ],
              "title": "Side",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Evidence Preview",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/label-learning": {
      "post": {
        "operationId": "post_label_learning_apply_changes__change_id__label_learning_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelLearningApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelLearningApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Label Learning Apply",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/label-learning/preview": {
      "post": {
        "operationId": "post_label_learning_preview_changes__change_id__label_learning_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelLearningPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Label Learning Preview",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
    "/changes/{change_id}/views": {
      "patch": {
        "operationId": "patch_change_view_changes__change_id__views_patch",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Change View",
        "tags": [
          "changes",
          "changes"
        ]
      }
    },
...
  }
}